Fix:解决因int引起的BUG问题

+++ src/HupunClient.php
@@ -341,6 +341,6 @@ class HupunClient
     public function getMillisecond()
     {
         list($t1, $t2) = explode(' ', microtime());
-        return (int)sprintf('%.0f', (floatval($t1) + floatval($t2)) * 1000);
+        return sprintf('%.0f', (floatval($t1) + floatval($t2)) * 1000);
     }    
 }
